refactor(trpc): extract createStripeSession into routers/billing
Continues p2-22 by carving the Stripe billing procedure out of
`src/trpc/index.ts` into `src/trpc/routers/billing.ts`. The dynamic
imports of `@/lib/stripe` and `@/config/stripe` are preserved exactly so
Stripe is still kept off the cold-path bundle. Behavior identical (same
billing-portal-vs-checkout branch, same metadata, same success/cancel
URLs).

Also drops the now-unused `absoluteUrl` import from index.ts since the
billing route was its only caller in this module.

Verification: 28/28 tRPC tests pass (billing + admin + index + sharing
+ ownership).

Refs: p2-22-split-trpc
